+++ test/eventsubscriber_test.dart
@@ -1,15 +1,5 @@
 import 'package:test/test.dart';
 
 void main() {
-  group('EventSubscriber', () {
-    // Awesome awesome;
-
-    // setUp(() {
-    //   awesome = Awesome();
-    // });
-
-    test('First Test', () {
-      // expect(awesome.isAwesome, isTrue);
-    });
-  });
+  test('See the test/s in the Example application included', () {});
 }